+++ Assets/Plugins/Behaviours/Editor/CopyPaste/CopyPasteContext.cs
@@ -1,0 +1,5 @@
+﻿namespace Jackey.Behaviours.Editor.CopyPaste {
+	public enum CopyPasteContext {
+		BT,
+	}
+}