+++ Assets/Plugins/Behaviours/Editor/PropertyDrawers/BehaviourConditionGroupPropertyDrawer.cs
@@ -1,4 +1,6 @@
-﻿using System.Reflection;
+﻿using System;
+using System.Linq;
+using System.Reflection;
 using Jackey.Behaviours.Attributes;
 using Jackey.Behaviours.Core.Conditions;
 using UnityEditor;
@@ -8,7 +10,10 @@ using UnityEngine.UIElements;
 namespace Jackey.Behaviours.Editor.PropertyDrawers {
 	[CustomPropertyDrawer(typeof(BehaviourConditionGroup))]
 	public class BehaviourConditionGroupPropertyDrawer : ManagedListPropertyDrawer<BehaviourCondition> {
+		internal static readonly Type[] s_conditionTypes = TypeCache.GetTypesDerivedFrom<BehaviourCondition>().Where(type => !type.IsAbstract).ToArray();
+
 		protected override string CreateButtonText => "Add Condition";
+		protected override Type[] CreateTypes => s_conditionTypes;
 
 		public override VisualElement CreatePropertyGUI(SerializedProperty property) {
 			VisualElement rootVisualElement = new VisualElement();